+++ src/Entities/Games/Game.php
@@ -22,10 +22,12 @@ use Longman\TelegramBot\Entities\PhotoSize;
  *
  * @link https://core.telegram.org/bots/api#game
  *
- * @method string    getTitle()        Title of the game
- * @method string    getDescription()  Description of the game
- * @method string    getText()         Optional. Brief description of the game or high scores included in the game message. Can be automatically edited to include current high scores for the game when the bot calls setGameScore, or manually edited using editMessageText. 0-4096 characters.
- * @method Animation getAnimation()    Optional. Animation that will be displayed in the game message in chats. Upload via BotFather
+ * @method string          getTitle()        Title of the game
+ * @method string          getDescription()  Description of the game
+ * @method PhotoSize[]     getPhoto()        Photo that will be displayed in the game message in chats.
+ * @method string          getText()         Optional. Brief description of the game or high scores included in the game message. Can be automatically edited to include current high scores for the game when the bot calls setGameScore, or manually edited using editMessageText. 0-4096 characters.
+ * @method MessageEntity[] getTextEntities() Optional. Special entities that appear in text, such as usernames, URLs, bot commands, etc.
+ * @method Animation       getAnimation()    Optional. Animation that will be displayed in the game message in chats. Upload via BotFather
  **/
 class Game extends Entity
 {
@@ -35,39 +37,9 @@ class Game extends Entity
     protected function subEntities()
     {
         return [
-            'photo'         => PhotoSize::class,
-            'text_entities' => MessageEntity::class,
+            'photo'         => [PhotoSize::class],
+            'text_entities' => [MessageEntity::class],
             'animation'     => Animation::class,
         ];
-    }
-
-    /**
-     * Photo that will be displayed in the game message in chats.
-     *
-     * This method overrides the default getPhoto method
-     * and returns a nice array of PhotoSize objects.
-     *
-     * @return null|PhotoSize[]
-     */
-    public function getPhoto()
-    {
-        $pretty_array = $this->makePrettyObjectArray(PhotoSize::class, 'photo');
-
-        return empty($pretty_array) ? null : $pretty_array;
-    }
-
-    /**
-     * Optional. Special entities that appear in text, such as usernames, URLs, bot commands, etc.
-     *
-     * This method overrides the default getTextEntities method
-     * and returns a nice array of MessageEntity objects.
-     *
-     * @return null|MessageEntity[]
-     */
-    public function getTextEntities()
-    {
-        $pretty_array = $this->makePrettyObjectArray(MessageEntity::class, 'text_entities');
-
-        return empty($pretty_array) ? null : $pretty_array;
     }
 }