+++ src/Entities/Payments/PreCheckoutQuery.php
@@ -35,7 +35,7 @@ class PreCheckoutQuery extends Entity
     /**
      * {@inheritdoc}
      */
-    public function subEntities()
+    protected function subEntities()
     {
         return [
             'from'       => User::class,