+++ src/Entities/Payments/ShippingQuery.php
@@ -32,7 +32,7 @@ class ShippingQuery extends Entity
     /**
      * {@inheritdoc}
      */
-    public function subEntities()
+    protected function subEntities()
     {
         return [
             'from'             => User::class,